platform-ui/requirements.txt
requests
beautifulsoup4
pillow
openai